+++ java-sample/src/main/java/io/snabble/testapp/PayoneInputActivity.java
@@ -1,0 +1,14 @@
+package io.snabble.testapp;
+
+import androidx.fragment.app.Fragment;
+
+import io.snabble.sdk.ui.integration.PayoneInputFragment;
+
+public class PayoneInputActivity extends BaseActivity {
+    @Override
+    public Fragment onCreateFragment() {
+        PayoneInputFragment fragment =  new PayoneInputFragment();
+        fragment.setArguments(getIntent().getBundleExtra("args"));
+        return fragment;
+    }
+}